requirements.txt
alembic==1.13.2
aniso8601==9.0.1
annotated-types==0.7.0
APScheduler==3.10.4
apturl==0.5.2
asn1crypto==1.5.1
atpublic==4.1.0
attrs==23.2.0
bcrypt==3.1.7
blinker==1.8.2
Brlapi==0.7.0
certifi==2024.6.2
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==2.2.1
colorama==0.4.3
cryptography==42.0.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distro==1.4.0
distro-info==0.23+ubuntu1.1
duplicity==0.8.12.0
entrypoints==0.3
Faker==25.8.0
fasteners==0.14.1
filelock==3.15.1
flasgger==0.9.7.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
flask-swagger-ui==4.11.1
future==0.18.2
greenlet==3.0.3
httplib2==0.14.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.3.5
MarkupSafe==2.1.5
mistune==3.0.2
monotonic==1.5
oauthlib==3.1.0
olefile==0.46
packaging==24.1
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==4.2.2
protobuf==3.6.1
psycopg2-binary==2.9.9
pycairo==1.16.2
pycparser==2.22
pycups==1.9.73
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.3.1
PyGObject==3.36.0
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==24.1.0
pyRFC3339==1.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.9.0.post0
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.1
pytz==2024.1
pyxdg==0.26
PyYAML==6.0.1
referencing==0.35.1
reportlab==3.5.34
requests==2.32.3
requests-unixsocket==0.2.0
rpds-py==0.18.1
SecretStorage==2.3.1
simplejson==3.16.0
six==1.16.0
snowflake==0.9.0
snowflake-connector-python==3.10.1
snowflake-id==1.0.2
snowflake-snowpark-python==1.18.0
snowflake._legacy==0.9.0
snowflake.core==0.9.0
sortedcontainers==2.4.0
SQLAlchemy==2.0.31
SQLAlchemy-serializer==1.4.12
systemd-python==234
tomlkit==0.12.5
typing_extensions==4.12.1
tzlocal==5.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
unattended-upgrades==0.1
urllib3==2.2.1
usb-creator==0.3.7
wadllib==1.3.3
Werkzeug==3.0.3
xkit==0.0.0
zipp==3.19.1
